# api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: UVP machine-client API
  version: 1.0.0
  description: >-
    Server-to-server machine-client boundary. The access key identifies a
    client and the server resolves its owner and enabled scopes. The external
    business API never returns SK; controlled administrative creation or
    rotation displays it once for secure server-side delivery.
servers:
  - url: /
...
security:
  - uvpHmac: []
paths:
...
  /openapi/v1/devices/{deviceId}/channels:
    get:
      operationId: listChannels
      summary: List channels of one owned device
      x-uvp-scope: channel:list
      x-uvp-enabled: true
      security:
        - uvpHmac: []
      parameters:
        - $ref: '#/components/parameters/SignVersion'
        - $ref: '#/components/parameters/AccessKey'
        - $ref: '#/components/parameters/Timestamp'
        - $ref: '#/components/parameters/Nonce'
        - $ref: '#/components/parameters/Signature'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Keyword'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Channel list
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelListResponse'
        '400': *invalidRequestResponse
        '401': *authenticationFailedResponse
        '403': *capabilityDeniedResponse
        '404': *resourceNotFoundResponse
        '405': *methodNotAllowedResponse
        '429': *rateLimitedResponse
        '503': *serviceUnavailableResponse
  /openapi/v1/devices/{deviceId}/channels/{channelId}:
    get:
      operationId: getChannel
      summary: Get one channel of one owned device
      x-uvp-scope: channel:detail
      x-uvp-enabled: true
      security:
        - uvpHmac: []
      parameters:
        - $ref: '#/components/parameters/SignVersion'
        - $ref: '#/components/parameters/AccessKey'
        - $ref: '#/components/parameters/Timestamp'
        - $ref: '#/components/parameters/Nonce'
        - $ref: '#/components/parameters/Signature'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Channel
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '400': *invalidRequestResponse
        '401': *authenticationFailedResponse
        '403': *capabilityDeniedResponse
        '404': *resourceNotFoundResponse
        '405': *methodNotAllowedResponse
        '429': *rateLimitedResponse
        '503': *serviceUnavailableResponse
  /openapi/v1/devices/{deviceId}/channels/{channelId}/status:
    get:
      operationId: getChannelStatus
      summary: Get the status of one owned channel
      x-uvp-scope: channel:status
      x-uvp-enabled: true
      security:
        - uvpHmac: []
      parameters:
        - $ref: '#/components/parameters/SignVersion'
        - $ref: '#/components/parameters/AccessKey'
        - $ref: '#/components/parameters/Timestamp'
        - $ref: '#/components/parameters/Nonce'
        - $ref: '#/components/parameters/Signature'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Channel status
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelStatusResponse'
        '400': *invalidRequestResponse
        '401': *authenticationFailedResponse
        '403': *capabilityDeniedResponse
        '404': *resourceNotFoundResponse
        '405': *methodNotAllowedResponse
        '429': *rateLimitedResponse
        '503': *serviceUnavailableResponse
...
components:
  securitySchemes:
    uvpHmac:
      type: apiKey
      in: header
      name: X-UVP-Signature
      description: HMAC-SHA256 signature produced with the server-side secret key.
  parameters:
    SignVersion:
      name: X-UVP-Sign-Version
      in: header
      required: true
      description: Signature version. Only v1 is accepted.
      schema:
        type: string
        enum: ['1']
    AccessKey:
      name: X-UVP-Access-Key
      in: header
      required: true
      description: Public access key; the matching secret key remains server-side.
      schema:
        type: string
        pattern: '^uvp_[0-9a-f]{32}$'
    Timestamp:
      name: X-UVP-Timestamp
      in: header
      required: true
      description: Unix timestamp in seconds, decimal without leading zero, at most 19 digits and no greater than 9223372036854775807.
      schema:
        type: string
        pattern: '^(0|[1-9][0-9]{0,18})$'
    Nonce:
      name: X-UVP-Nonce
      in: header
      required: true
      description: Single-use request nonce; duplicate and comma-joined values are rejected.
      schema:
        type: string
        pattern: '^[0-9a-f]{32}$'
    Signature:
      name: X-UVP-Signature
      in: header
      required: true
      description: Lowercase hexadecimal HMAC-SHA256 signature over the canonical request.
      schema:
        type: string
        pattern: '^[0-9a-f]{64}$'
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9]{20}$'
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9]{20}$'
    Page:
      name: page
      in: query
      required: false
      description: Positive one-based page number; defaults to 1 when omitted; an explicit empty value is rejected.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Positive page size, at most 100; defaults to 20 when omitted; an explicit empty value is rejected.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    Keyword:
      name: keyword
      in: query
      required: false
      description: Valid UTF-8 search text, limited to 100 Unicode characters; an empty value is valid.
      schema:
        type: string
        maxLength: 100
    Status:
      name: status
      in: query
      required: false
      description: Optional device or channel status filter; an explicit empty value is rejected.
      schema:
        type: string
        enum: [online, offline, unknown]
  schemas:
...
    Channel:
      type: object
      additionalProperties: false
      required: [deviceId, channelId, name, alias, manufacturer, model, status, ptzType]
      properties:
        deviceId: {type: string}
        channelId: {type: string}
        name: {type: string}
        alias: {type: string}
        manufacturer: {type: string}
        model: {type: string}
        status:
          type: string
          enum: [online, offline, unknown]
        ptzType: {type: integer, format: int32}
    ChannelPage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Channel'}
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, format: int64, minimum: 0}
    ChannelStatus:
      type: object
      additionalProperties: false
      required: [deviceId, channelId, status]
      properties:
        deviceId: {type: string}
        channelId: {type: string}
        status:
          type: string
          enum: [online, offline, unknown]
...
    ChannelListResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId, data]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, enum: [success]}
        requestId: {type: string}
        data: {$ref: '#/components/schemas/ChannelPage'}
    ChannelResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId, data]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, enum: [success]}
        requestId: {type: string}
        data: {$ref: '#/components/schemas/Channel'}
    ChannelStatusResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId, data]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, enum: [success]}
        requestId: {type: string}
        data: {$ref: '#/components/schemas/ChannelStatus'}
...